requirements.txt
yt-dlp
faster-whisper
openai
python-dotenv
pytesseract
pillow
